+++ pkg/appmanager/healthMonitors.go
@@ -24,7 +24,7 @@ import (
 
 	log "github.com/F5Networks/k8s-bigip-ctlr/pkg/vlogger"
 
-	"k8s.io/api/extensions/v1beta1"
+	"k8s.io/api/networking/v1beta1"
 )
 
 func (appMgr *Manager) assignHealthMonitorsByPath(